+++ apps/sdp-api/Dockerfile.dockerignore
@@ -20,6 +20,7 @@ packages/*
 !packages/sdp-types
 !packages/sdp-env-config
 !packages/sdp-rpc
+!packages/sdp-custody
 docs
 infra
 scripts
