+++ apps/sdp-api/package.json
@@ -42,6 +42,7 @@
     "@hono/node-server": "^2.0.4",
     "@react-email/components": "1.0.12",
     "@react-email/render": "2.0.8",
+    "@sdp/custody": "workspace:*",
     "@sdp/env-config": "workspace:*",
     "@sdp/rpc": "workspace:*",
     "@sdp/types": "workspace:*",
