+++ apps/sdp-api/src/lib/redaction.ts
@@ -1,101 +1,1 @@
-const REDACTED = "[REDACTED]";
-
-const SENSITIVE_JSON_FIELD_PATTERN =
-  /(["'])(app[-_ ]?secret|api[-_ ]?secret|api[-_ ]?key|client[-_ ]?secret|wallet[-_ ]?secret|signing[-_ ]?secret|private[-_ ]?key|secret[-_ ]?key|access[-_ ]?token|refresh[-_ ]?token|id[-_ ]?token|authorization|password|pem|token|secret|credential)\1\s*:\s*(["'])(.*?)\3/gi;
-const SENSITIVE_ASSIGNMENT_PATTERN =
-  /\b(app[-_ ]?secret|api[-_ ]?secret|api[-_ ]?key|client[-_ ]?secret|wallet[-_ ]?secret|signing[-_ ]?secret|private[-_ ]?key|secret[-_ ]?key|access[-_ ]?token|refresh[-_ ]?token|id[-_ ]?token|authorization|password|pem|token|secret|credential)\b(\s*[:=]\s*)[^,\s}]+/gi;
-const PRIVATE_KEY_PATTERN =
-  /-----BEGIN [^-]*PRIVATE KEY-----[\s\S]*?-----END [^-]*PRIVATE KEY-----/g;
-const AUTH_HEADER_PATTERN = /\b(Bearer|Basic)\s+([A-Za-z0-9._~+/=-]+)/gi;
-
-export function redactCredentialString(value: string): string {
-  return value
-    .replace(PRIVATE_KEY_PATTERN, REDACTED)
-    .replace(AUTH_HEADER_PATTERN, (match, scheme: string, token: string) =>
-      isLikelyAuthToken(scheme, token) ? `${scheme} ${REDACTED}` : match
-    )
-    .replace(
-      SENSITIVE_JSON_FIELD_PATTERN,
-      (_match, quote: string, key: string, valueQuote: string) =>
-        `${quote}${key}${quote}:${valueQuote}${REDACTED}${valueQuote}`
-    )
-    .replace(
-      SENSITIVE_ASSIGNMENT_PATTERN,
-      (_match, key: string, separator: string) => `${key}${separator}${REDACTED}`
-    );
-}
-
-function isLikelyAuthToken(scheme: string, token: string): boolean {
-  const minLength = scheme.toLowerCase() === "basic" ? 12 : 16;
-  return token.length >= minLength || /[0-9._~+/=-]/.test(token);
-}
-
-export function redactCredentialSecrets<T>(value: T): T {
-  return redactValue(value, new WeakSet<object>()) as T;
-}
-
-function redactValue(value: unknown, seen: WeakSet<object>): unknown {
-  if (typeof value === "string") {
-    return redactCredentialString(value);
-  }
-
-  if (value === null || value === undefined || typeof value !== "object") {
-    return value;
-  }
-
-  if (seen.has(value)) {
-    return "[Circular]";
-  }
-  seen.add(value);
-
-  if (value instanceof Error) {
-    return {
-      name: value.name,
-      message: redactCredentialString(value.message),
-      ...(value.stack ? { stack: redactCredentialString(value.stack) } : {}),
-      ...("cause" in value ? { cause: redactValue(value.cause, seen) } : {}),
-    };
-  }
-
-  if (value instanceof Date) {
-    return value.toISOString();
-  }
-
-  if (value instanceof URL) {
-    return value.toString();
-  }
-
-  if (value instanceof ArrayBuffer || ArrayBuffer.isView(value)) {
-    return REDACTED;
-  }
-
-  if (Array.isArray(value)) {
-    return value.map((item) => redactValue(item, seen));
-  }
-
-  const result: Record<string, unknown> = {};
-  for (const [key, item] of Object.entries(value)) {
-    result[key] = isSensitiveCredentialKey(key) ? REDACTED : redactValue(item, seen);
-  }
-  return result;
-}
-
-function isSensitiveCredentialKey(key: string): boolean {
-  const normalized = key.replace(/[^a-z0-9]/gi, "").toLowerCase();
-  return (
-    normalized === "secret" ||
-    normalized === "credential" ||
-    normalized === "credentials" ||
-    normalized === "apikey" ||
-    normalized === "authorization" ||
-    normalized === "password" ||
-    normalized === "pem" ||
-    normalized === "token" ||
-    normalized.endsWith("secret") ||
-    normalized.endsWith("password") ||
-    normalized.endsWith("token") ||
-    normalized.endsWith("pem") ||
-    normalized.includes("privatekey") ||
-    normalized.includes("secretpayload")
-  );
-}
+export { redactCredentialSecrets, redactCredentialString } from "@sdp/custody";
