+++ packages/sdp-custody/package.json
@@ -1,0 +1,47 @@
+{
+  "name": "@sdp/custody",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts",
+      "default": "./src/index.ts"
+    },
+    "./signing": {
+      "types": "./src/signing.ts",
+      "import": "./src/signing.ts",
+      "default": "./src/signing.ts"
+    },
+    "./keychain": {
+      "types": "./src/keychain/index.ts",
+      "import": "./src/keychain/index.ts",
+      "default": "./src/keychain/index.ts"
+    },
+    "./dfns": {
+      "types": "./src/dfns/index.ts",
+      "import": "./src/dfns/index.ts",
+      "default": "./src/dfns/index.ts"
+    }
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "lint": "biome check src/"
+  },
+  "dependencies": {
+    "@sdp/types": "workspace:*",
+    "@solana/addresses": "catalog:",
+    "@solana/codecs": "catalog:",
+    "@solana/keychain-cdp": "1.3.0",
+    "@solana/keychain-core": "1.3.0",
+    "@solana/keychain-fireblocks": "1.3.0",
+    "@solana/keychain-para": "1.3.0",
+    "@solana/keychain-privy": "1.3.0",
+    "@solana/keychain-turnkey": "1.3.0",
+    "@solana/keychain-utila": "1.3.0",
+    "@solana/kit": "catalog:",
+    "@solana/signers": "catalog:",
+    "@solana/transactions": "catalog:"
+  }
+}
